+++ arch/arm/dev/ih_at91irq0.c
@@ -33,6 +33,9 @@
 
 /*
  * $Log$
+ * Revision 1.5  2008/08/11 06:59:09  haraldkipp
+ * BSD types replaced by stdint types (feature request #1282721).
+ *
  * Revision 1.4  2006/06/28 17:10:15  haraldkipp
  * Include more general header file for ARM.
  *
@@ -99,8 +102,8 @@ void Interrupt0Entry(void)
 static int Interrupt0Ctl(int cmd, void *param)
 {
     int rc = 0;
-    u_int *ival = (u_int *)param;
-    int enabled = inr(AIC_IMR) & _BV(IRQ0_ID);
+    unsigned int *ival = (unsigned int *)param;
+    int_fast8_t enabled = inr(AIC_IMR) & _BV(IRQ0_ID);
 
     /* Disable interrupt. */
     if (enabled) {
@@ -132,7 +135,7 @@ static int Interrupt0Ctl(int cmd, void *param)
         break;
     case NUT_IRQCTL_GETMODE:
         {
-            u_int val = inr(AIC_SMR(IRQ0_ID)) & AIC_SRCTYPE;
+            unsigned int val = inr(AIC_SMR(IRQ0_ID)) & AIC_SRCTYPE;
             if (val == AIC_SRCTYPE_EXT_LOW_LEVEL) {
                 *ival = NUT_IRQMODE_LOWLEVEL;
             } else if (val == AIC_SRCTYPE_EXT_HIGH_LEVEL) {
@@ -165,7 +168,7 @@ static int Interrupt0Ctl(int cmd, void *param)
         break;
 #ifdef NUT_PERFMON
     case NUT_IRQCTL_GETCOUNT:
-        *ival = (u_int)sig_INTERRUPT0.ir_count;
+        *ival = (unsigned int)sig_INTERRUPT0.ir_count;
         sig_INTERRUPT0.ir_count = 0;
         break;
 #endif
