+++ arch/arm/dev/ih_at91irq1.c
@@ -33,6 +33,9 @@
 
 /*
  * $Log$
+ * Revision 1.4  2008/08/11 06:59:10  haraldkipp
+ * BSD types replaced by stdint types (feature request #1282721).
+ *
  * Revision 1.3  2006/06/28 17:10:15  haraldkipp
  * Include more general header file for ARM.
  *
@@ -96,8 +99,8 @@ void Interrupt1Entry(void)
 static int Interrupt1Ctl(int cmd, void *param)
 {
     int rc = 0;
-    u_int *ival = (u_int *)param;
-    int enabled = inr(AIC_IMR) & _BV(IRQ1_ID);
+    unsigned int *ival = (unsigned int *)param;
+    int_fast8_t enabled = inr(AIC_IMR) & _BV(IRQ1_ID);
 
     /* Disable interrupt. */
     if (enabled) {
@@ -129,7 +132,7 @@ static int Interrupt1Ctl(int cmd, void *param)
         break;
     case NUT_IRQCTL_GETMODE:
         {
-            u_int val = inr(AIC_SMR(IRQ1_ID)) & AIC_SRCTYPE;
+            unsigned int val = inr(AIC_SMR(IRQ1_ID)) & AIC_SRCTYPE;
             if (val == AIC_SRCTYPE_EXT_LOW_LEVEL) {
                 *ival = NUT_IRQMODE_LOWLEVEL;
             } else if (val == AIC_SRCTYPE_EXT_HIGH_LEVEL) {
@@ -162,7 +165,7 @@ static int Interrupt1Ctl(int cmd, void *param)
         break;
 #ifdef NUT_PERFMON
     case NUT_IRQCTL_GETCOUNT:
-        *ival = (u_int)sig_INTERRUPT1.ir_count;
+        *ival = (unsigned int)sig_INTERRUPT1.ir_count;
         sig_INTERRUPT1.ir_count = 0;
         break;
 #endif
