+++ arch/arm/dev/ih_at91spi0.c
@@ -33,6 +33,9 @@
 
 /*
  * $Log$
+ * Revision 1.3  2008/08/11 06:59:10  haraldkipp
+ * BSD types replaced by stdint types (feature request #1282721).
+ *
  * Revision 1.2  2008/07/26 09:43:01  haraldkipp
  * Added support for retrieving and setting the interrupt mode.
  *
@@ -95,8 +98,8 @@ void SerialPeripheral0IrqEntry(void)
 static int SerialPeripheral0IrqCtl(int cmd, void *param)
 {
     int rc = 0;
-    u_int *ival = (u_int *) param;
-    int enabled = inr(AIC_IMR) & _BV(SPI0_ID);
+    unsigned int *ival = (unsigned int *) param;
+    int_fast8_t enabled = inr(AIC_IMR) & _BV(SPI0_ID);
 
     /* Disable interrupt. */
     if (enabled) {
@@ -127,7 +130,7 @@ static int SerialPeripheral0IrqCtl(int cmd, void *param)
         break;
     case NUT_IRQCTL_GETMODE:
         {
-            u_int val = inr(AIC_SMR(SPI0_ID)) & AIC_SRCTYPE;
+            unsigned int val = inr(AIC_SMR(SPI0_ID)) & AIC_SRCTYPE;
             if (val == AIC_SRCTYPE_INT_LEVEL_SENSITIVE || val == AIC_SRCTYPE_EXT_HIGH_LEVEL) {
                 *ival = NUT_IRQMODE_LEVEL;
             } else  {
@@ -152,7 +155,7 @@ static int SerialPeripheral0IrqCtl(int cmd, void *param)
         break;
 #ifdef NUT_PERFMON
     case NUT_IRQCTL_GETCOUNT:
-        *ival = (u_int) sig_SPI0.ir_count;
+        *ival = (unsigned int) sig_SPI0.ir_count;
         sig_SPI0.ir_count = 0;
         break;
 #endif
